refactor(log): promote the async log lister into log/_file.py (#4622)
Move `list_eval_logs_async` and its `async_filesystem`/walk helpers out of
`_view/common.py` into `log/_file.py`, so the public API's listing is
genuinely async (s3 fast path, async fsspec walk) rather than blocking the
event loop on remote log dirs. `s3fs`/`botocore` are imported lazily so they
don't become import-time dependencies of `inspect_ai.log`.

Also:
- keep recover's re-read lazy under asyncio when not overwriting
- switch the remaining sync `read_eval_log()` verification reads in
  test_recover_api.py to `read_eval_log_async()` so the file is trio-clean
- add direct `list_eval_logs_async` tests (no-filter, filter, missing dir)
- CHANGELOG entry
