+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Memory tool: Canonicalize `/memories` paths so equivalent spellings map to one file rather than several, and add an `instance` parameter to `memory()` for independent per-instance memory stores.
 - Registry: Add a `validation_predicate` type for Scout extensions.
 - Logging: Realtime streaming sample writes now normalize unserializable store values the same way as the standard log write path, instead of raising during serialization. (#4120)
+- Logging: New `list_eval_logs_async()`, and eval log recovery (`recover_eval_log_async()`, `recoverable_eval_logs_async()`) can now be called from a trio event loop. (#4622)
 - Tracing: trace-file reads (`inspect trace` commands and `inspect ctl process anomalies`) now skip truncated, corrupt, or unrecognized records — e.g. from a hard-killed process or a newer inspect version — instead of failing entirely.
 - Windows: process liveness checks (used by `inspect ctl` discovery and pid targeting) no longer risk sending a console Ctrl+C to the probed process or misreporting live processes as dead.
 - Approval: Load policy configuration files from `file://` URIs as well as plain local paths.
