+++ src/inspect_ai/log/_recover/_api.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Iterator
 
-from inspect_ai._util._async import run_coroutine
+from inspect_ai._util._async import current_async_backend, run_coroutine
 from inspect_ai._util.file import dirname, filesystem
 from inspect_ai.log._file import (
     EvalLogInfo,
@@ -212,14 +212,17 @@ async def recover_eval_log_async(
         fs.rm(final_output)
         fs.mv(write_output, final_output)
 
-    # Re-read the written log through the async reader before returning. The
-    # log produced by write_recovered_eval_log() carries a LazyList of samples
-    # whose first access calls the sync read_eval_log(), which raises under a
-    # trio async context; reading eagerly here resolves samples to a plain list
-    # so the returned log stays usable end to end under both asyncio and trio
-    # (and, for overwrite mode, also repoints sample data at the final path).
-    final_path = final_output if final_output is not None else write_output
-    recovered_log = await read_eval_log_async(final_path)
+    # Re-read the written log through the async reader when we must. The log
+    # produced by write_recovered_eval_log() carries a LazyList of samples whose
+    # first access calls the sync read_eval_log(), which raises under a trio
+    # async context; reading eagerly resolves samples to a plain list so the
+    # returned log stays usable end to end under trio. Overwrite mode re-reads
+    # regardless, to repoint sample data at the final path. Under asyncio with
+    # no overwrite the samples stay lazy, so callers that only read the header
+    # (e.g. `inspect log recover`) don't pay a full deserialization.
+    if final_output is not None or current_async_backend() == "trio":
+        final_path = final_output if final_output is not None else write_output
+        recovered_log = await read_eval_log_async(final_path)
 
     # Cleanup buffer DB (only after successful write)
     if cleanup and recovery_data is not None and recovery_data.buffer is not None:
